feat: stats
refactor to better organize the Statistics folder
