+++ src/pages/TheCrucible/Statistics/index.ts
@@ -1,2 +1,2 @@
 export { default as Statisics } from "./Statistics";
-export { default as StatInput } from "./StatInput";
+export { default as StatInput } from "./components/StatInput/StatInput";
